upgrade multiple service + testcases + java docs

+++ src/main/java/com/engg/digitalorg/managers/UrlManager.java
@@ -12,30 +12,41 @@ import java.util.Date;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 
+/**
+ * The type Url manager.
+ */
 @Component
 public class UrlManager {
 
     @Autowired
     private UrlRepository urlRepository;
 
+    /**
+     * The Base conversion.
+     */
     @Autowired
     BaseConversion baseConversion;
 
 
+    /**
+     * Gets original url.
+     *
+     * @param shortUrl the short url
+     * @return the original url
+     */
     public String getOriginalUrl(String shortUrl) {
         Optional<Url> entity;
         try {
-            long id = baseConversion.decode(shortUrl);
+            int id = baseConversion.decode(shortUrl);
             entity = urlRepository.findById(id);
-            if(entity == null || entity.get() == null) {
+            if (entity == null || entity.get() == null) {
                 throw new EntityNotFoundException("There is no record with " + shortUrl);
             }
-            if (entity.get().getExpires_date() != null && entity.get().getExpires_date().before(new Date())){
+            if (entity.get().getExpires_date() != null && entity.get().getExpires_date().before(new Date())) {
                 urlRepository.delete(entity.get());
                 throw new EntityNotFoundException("Link expired!");
             }
-        }
-        catch (NoSuchElementException exception) {
+        } catch (NoSuchElementException exception) {
             throw new DigitalOrgException("There is no record with " + shortUrl);
         }
